Type corrections, Constants moved to config, IMU fix
Every variable type is corrected following the directives,
Some constant values are moved to the config.h.
The temperature measurement functionality is removed form the IMU class, because of no use.

+++ drone_controller/debug_in.cpp
@@ -132,7 +132,7 @@ debug_in& debug_in::operator>>(double &i) {
 		if (read_char < '0' || read_char > '9') return *this;
 	}
 	else return *this;
-	int temp = 10;
+	uint64_t temp = 10;
 	double temp2= read_char - '0';
 	temp2 = (double)temp2 / (double)temp;
 	if (negative) i -= temp2;
